feat: Add more conversion functionality
Also reorganize the existing task/mint stuff to make it more sensible
with new additions.

Rename the task package to convert, which matches the subcommand name
and is overall more descriptive of the provided functionality.

Change a YNAB method to a function and move to convert package since
that's where it's used and that package is all about parsing and
formatting stuff.