refactor(convert): Adjust csv parsing to accept stream
Update the readParse{servicename} funcs to allow more than one file at a
time. Add some simple detection for header or metadata rows because the
number of rows to skip can no longer be known ahead of time. It has to
accept a continuous stream of data.

The use case is to process many files at once, as if you piped data from
STDIN like this: "command < path/to/data/*.csv".